added LinuxDebian Installer for v1.1
With running this script, users of Linux Debian system, and probably also other Linux distribution and perhaps macOS as well, can easily set up the data structure Notando 1.1 requires. Users have to agree to the Notando Terms of Use as well as the Notando Data Protection Policy to proceed. After running, your system is ready to run the main script